requirements.txt
altgraph==0.17
click==7.1.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.23
pefile==2019.4.18
protobuf==3.15.5
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
pywin32-ctypes==0.2.0
six==1.15.0
Werkzeug==1.0.1
